Fixed: Don't die on movie refresh when collection has been deleted from TMDB
Fixes #8664

+++ src/NzbDrone.Core/Movies/Collections/MovieCollectionService.cs
@@ -122,6 +122,11 @@ namespace NzbDrone.Core.Movies.Collections
 
                 var collection = FindByTmdbId(collectionTmdbId);
 
+                if (collection == null)
+                {
+                    continue;
+                }
+
                 _repo.Delete(collection.Id);
 
                 _eventAggregator.PublishEvent(new CollectionDeletedEvent(collection));
